libft: fixed strlen

+++ src/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: Xifeng <xifeng@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 17:07:41 by Xifeng            #+#    #+#             */
-/*   Updated: 2024/11/01 17:23:15 by Xifeng           ###   ########.fr       */
+/*   Updated: 2024/11/01 17:34:28 by Xifeng           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
